Refatoração do código
