--- app/src/main/java/br/com/alura/agenda/dao/AlunoDAO.java
@@ -1,54 +1,0 @@
-package br.com.alura.agenda.dao;
-
-import androidx.annotation.Nullable;
-
-import java.util.ArrayList;
-import java.util.List;
-
-import br.com.alura.agenda.models.Aluno;
-
-public class AlunoDAO {
-
-    private final static List<Aluno> alunos = new ArrayList<>();
-    private static int contadorDeIds = 1;
-
-    public void salva(Aluno aluno) {
-        aluno.setId(contadorDeIds);
-        alunos.add(aluno);
-        atualizaIds();
-    }
-
-    private static void atualizaIds() {
-        contadorDeIds++;
-    }
-
-    public void edita(Aluno aluno) {
-        Aluno alunoEncontrado = buscaAlunoPeloId(aluno);
-        if (alunoEncontrado != null) {
-            int posicaoDoAluno = alunos.indexOf(alunoEncontrado);
-            alunos.set(posicaoDoAluno, aluno);
-        }
-    }
-
-    @Nullable
-    private Aluno buscaAlunoPeloId(Aluno aluno) {
-        for (Aluno a :
-                alunos) {
-            if (a.getId() == aluno.getId()) {
-                return a;
-            }
-        }
-        return null;
-    }
-
-    public List<Aluno> todos() {
-        return new ArrayList<>(alunos);
-    }
-
-    public void remove(Aluno aluno) {
-        Aluno alunoDevolvido = buscaAlunoPeloId(aluno);
-        if (alunoDevolvido != null) {
-            alunos.remove(alunoDevolvido);
-        }
-    }
-}
