+++ app/src/main/java/br/com/alura/agenda/database/AgendaDatabase.java
@@ -1,5 +1,7 @@
 package br.com.alura.agenda.database;
 
+import static br.com.alura.agenda.database.AgendaMigrations.TODAS_MIGRATIONS;
+
 import android.content.Context;
 
 import androidx.annotation.NonNull;
@@ -26,37 +28,7 @@ public abstract class AgendaDatabase extends RoomDatabase {
         return Room
                 .databaseBuilder(context, AgendaDatabase.class, NOME_BANCO_DE_DADOS)
                 .allowMainThreadQueries()
-                .addMigrations(new Migration(1, 2) {
-                    @Override
-                    public void migrate(@NonNull SupportSQLiteDatabase database) {
-                        database.execSQL("ALTER TABLE Aluno ADD COLUMN sobrenome TEXT");
-                    }
-                }, new Migration(2, 3) {
-                    @Override
-                    public void migrate(@NonNull SupportSQLiteDatabase database) {
-                        // Criar nova tabela com as informações desejadas
-                        database.execSQL("CREATE TABLE IF NOT EXISTS `Aluno_novo` " +
-                                "(`id` INTEGER PRIMARY KEY AUTOINCREMENT NOT NULL, " +
-                                "`nome` TEXT, " +
-                                "`telefone` TEXT, " +
-                                "`email` TEXT)");
-
-                        // Copiar dados da tabela antiga para a nova
-                        database.execSQL("INSERT INTO Aluno_novo (id, nome, telefone, email) " +
-                                "SELECT id, nome, telefone, email FROM Aluno");
-
-                        // Remove tabela antiga
-                        database.execSQL("DROP TABLE Aluno");
-
-                        // Renomear tabela nova com o nome da tabela antiga
-                        database.execSQL("ALTER TABLE Aluno_novo RENAME TO Aluno");
-                    }
-                }, new Migration(3, 4) {
-                    @Override
-                    public void migrate(@NonNull SupportSQLiteDatabase database) {
-                        database.execSQL("ALTER TABLE Aluno ADD COLUMN momentoDeCadastro INTEGER");
-                    }
-                })
+                .addMigrations(TODAS_MIGRATIONS)
                 .build();
     }
 }
