+++ app/src/main/java/br/com/alura/agenda/models/Aluno.java
@@ -27,22 +27,6 @@ public class Aluno implements Serializable {
     public void setMomentoDeCadastro(Calendar momentoDeCadastro) {
         this.momentoDeCadastro = momentoDeCadastro;
     }
-    
-    public String dataFormatada() {
-        SimpleDateFormat formatador = new SimpleDateFormat("dd/MM/yyyy");
-        return formatador.format(momentoDeCadastro.getTime());
-    }
-
-    @Ignore
-    public Aluno(String nome, String telefone, String email) {
-        this.nome = nome;
-        this.telefone = telefone;
-        this.email = email;
-    }
-
-    public Aluno() {
-
-    }
 
     public int getId() {
         return id;
@@ -84,9 +68,5 @@ public class Aluno implements Serializable {
 
     public boolean temIdValido() {
         return id > 0;
-    }
-
-    public String getNomeCompleto() {
-        return nome;
     }
 }
